// src/modules/auth/dtos/auth.schema.ts
import { z } from "zod";

export const registerSchema = z.object({

  email: z.string().email(),

  password: z.string().min(6),

  name: z.string().min(2),

  phone: z.string().optional()

});

export const loginSchema = z.object({

  email: z.string().email(),

  password: z.string()

});

export type RegisterInput = z.infer<typeof registerSchema>;
export type LoginInput = z.infer<typeof loginSchema>;